test-producer/requirements.txt
kafka-python
hdfs[avro]
